requirements.txt
#treamlit==1.11.1
#pandas==1.3.5
#numpy==1.22.0
#scipy==1.2.1
#matplotlib==3.3.1
#python-dotenv==0.15.0  
#pydeck==0.7.1
#protobuf = ">=3.6.0, !=3.11"
#python-dotenv==0.20.0

streamlit== 1.11.1
pandas==1.5.0
numpy==1.22.0
pydeck==0.7.1
python-dotenv==0.20.0